Commands are now have it's own file, see wiki for references EXEC-activator is not basis. Again Removed COMMAND activator Added PRECOMMAND activator Action MESSAGE now can send titles and subtitles Refactoring, removed some duplicated code Some optimizations More TODOs for future updates ( ;-; ) Started implementing RaContext

+++ core/src/main/java/me/fromgate/reactions/flags/FlagWalkSpeed.java
@@ -12,7 +12,7 @@ public class FlagWalkSpeed implements Flag {
 	public boolean checkFlag(Player player, String param) {
 		if (!Util.isInteger(param)) return false;
 		long walkSpeed = Math.round(player.getWalkSpeed() * 10);
-		Variables.setTempVar("walkspeed", Integer.toString((int) walkSpeed));
+		Variables.setTempVar("walkspeed", Long.toString(walkSpeed));
 		return walkSpeed >= Integer.parseInt(param);
 
 	}
